refactor(preference): use pinia store
