+++ app/src/components/Breadcrumb/index.ts
@@ -1,0 +1,3 @@
+import Breadcrumb from './Breadcrumb.vue'
+
+export default Breadcrumb
