+++ app/src/components/EnvGroupTabs/index.ts
@@ -1,0 +1,3 @@
+import EnvGroupTabs from './EnvGroupTabs.vue'
+
+export default EnvGroupTabs
