+++ app/src/components/EnvIndicator/index.ts
@@ -1,0 +1,3 @@
+import EnvIndicator from './EnvIndicator.vue'
+
+export default EnvIndicator
