+++ app/src/components/ICP/index.ts
@@ -1,0 +1,3 @@
+import ICP from './ICP.vue'
+
+export default ICP
