+++ app/src/components/Logo/index.ts
@@ -1,0 +1,3 @@
+import Logo from './Logo.vue'
+
+export default Logo
