+++ app/src/components/NginxControl/index.ts
@@ -1,0 +1,3 @@
+import NginxControl from './NginxControl.vue'
+
+export default NginxControl
