+++ app/src/components/NodeSelector/index.ts
@@ -1,0 +1,3 @@
+import NodeSelector from './NodeSelector.vue'
+
+export default NodeSelector
