+++ app/src/components/OTPInput/index.ts
@@ -1,0 +1,3 @@
+import OTPInput from './OTPInput.vue'
+
+export default OTPInput
