+++ app/src/components/PageHeader/index.ts
@@ -1,0 +1,3 @@
+import PageHeader from './PageHeader.vue'
+
+export default PageHeader
