+++ app/src/components/ReactiveFromNow/index.ts
@@ -1,0 +1,3 @@
+import ReactiveFromNow from './ReactiveFromNow.vue'
+
+export default ReactiveFromNow
