+++ app/src/components/SensitiveString/index.ts
@@ -1,0 +1,3 @@
+import SensitiveString from './SensitiveString.vue'
+
+export default SensitiveString
