+++ app/src/components/SetLanguage/index.ts
@@ -1,0 +1,3 @@
+import SetLanguage from './SetLanguage.vue'
+
+export default SetLanguage
