+++ app/src/components/SwitchAppearance/index.ts
@@ -1,0 +1,3 @@
+import SwitchAppearance from './SwitchAppearance.vue'
+
+export default SwitchAppearance
