+++ app/src/components/SystemRestore/index.ts
@@ -1,0 +1,3 @@
+import SystemRestoreContainer from './SystemRestoreContent.vue'
+
+export default SystemRestoreContainer
