+++ app/src/components/VPSwitch/index.ts
@@ -1,0 +1,3 @@
+import VPSwitch from './VPSwitch.vue'
+
+export default VPSwitch
