+++ app/src/views/preference/components/AuthSettings/TOTP.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { RecoveryCode } from '@/api/recovery'
 import otp from '@/api/otp'
-import OTPInput from '@/components/OTPInput/OTPInput.vue'
-import use2FAModal from '@/components/TwoFA/use2FAModal'
+import OTPInput from '@/components/OTPInput'
+import { use2FAModal } from '@/components/TwoFA'
 import { CheckCircleOutlined } from '@ant-design/icons-vue'
 import { UseClipboard } from '@vueuse/components'
 import { message } from 'ant-design-vue'
